requirements.txt
numpy==1.14.2
matplotlib==2.2.2
pandas==0.23.4
pandas-datareader==0.6.0
scikit-image==0.14.0
scikit-learn==0.19.0
scipy==0.19.1
Pillow==5.1.0

dash==0.21.1
dash-core-components==0.23.0
dash-html-components==0.11.0
dash-renderer==0.13.0
dask==0.18.2

Flask==0.12.2
Flask-Compress==1.4.0
Flask-Cors==3.0.2

gevent==1.2.1
greenlet==0.4.12

gunicorn==19.9.0
h5py==2.8.0
Jinja2==2.9.6

jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.0.0


nose==1.3.7
notebook==5.0.0

requests==2.14.2
requests-file==1.4.3
requests-ftp==0.3.1

tensorboard==1.10.0
tensorflow==1.10.0

opencv-python==4.0.0.21